Final push

+++ app-react/src/ClickableUserProfile.jsx
@@ -1,11 +1,23 @@
 import React, {useContext} from 'react';
 import {UserContext} from './UserContext';
-import GetUrl from "./Url.jsx";
+
 import axios from 'axios';
+
+/**
+ * Component representing a clickable user profile.
+ * @param {Object} props - The component props.
+ * @param {Object} props.user - The user object.
+ * @param {Function} props.switchToProfile - Function to switch to user profile.
+ * @param {Function} props.setUpToDate - Function to update user data.
+ * @returns {JSX.Element} - The rendered component.
+ */
 function ClickableUserProfile(props) {
 
     const currentUser = useContext(UserContext);
     
+    /**
+     * Approves the user if the current user is an admin.
+     */
     const ApproveUser = () => {
         if(!currentUser.is_admin){return}
         axios.patch(`/users/${props.user.id}`, {field : "approved", value : true, })
@@ -17,6 +29,9 @@ function ClickableUserProfile(props) {
             });
     }
     
+    /**
+     * Promotes the user to admin if the current user is an admin.
+     */
     const PromoteUser = () => {
         if(!currentUser.is_admin){return}
         axios.patch(`/users/${props.user.id}`, {field : "is_admin", value : true})
@@ -28,7 +43,11 @@ function ClickableUserProfile(props) {
             });
     }
     
+    /**
+     * Deletes the user if the current user is an admin.
+     */
     const DeleteUser = () => {
+        if(!currentUser.is_admin){return}
         axios.delete(`/users/${props.user.id}`)
             .then((response) => {
                 if(response.status === 200) props.setUpToDate(true);
